fix: handle 404s during Read() correctly
re: https://github.com/planetscale/terraform-provider-planetscale/issues/53

When a resource in the statefile has been deleted out of band, such as someone on the console or
using pscale cli deleting it, terraform should remove it from state and plan to recreate it.
However, we were treating 404s as fatal.

To solve this, the `Read()` method in each of our four resources now checks for 404 and removes the
resource from state so that terraform will offer to recreate the resource

This PR also adds initial acceptance test coverage for `database`, `password`, and `branch`
resources, including tests for the out-of-band deletion handling fixed by this PR. The `backup`
resource is currently not working and so tests for it will be added in later PRs.