update backup resource to match latest API
This change introduces a breaking change to rhe `planetscale_backup` resource. This is necessary to align with the latest API changes.

The `planetscale_backup` resource is a one-time resource that creates a backup of a database at a specific point in time. Any updates to the resource will result in a recreation anyway.

This is not the same as a backup schedule which produces multiple backups over a period of time. Backup schedules are not yet supported in the Terraform provider.

Prior to this change a `planetscale_backup` resource would be defined with retention values in the `backup_policy` hash, eg:

```hcl
resource "planetscale_backup" "test" {
  organization = "myorg"
  name         = "my-backup"
  database     = "mydb"
  branch       = "mybranch"
  backup_policy = {
    retention_unit  = "d"
    retention_value = 7.0
  }
}
```

Now they must be defined like:

```hcl
resource "planetscale_backup" "test" {
  organization    = "myorg"
  name            = "my-backup"
  database        = "mydb"
  branch          = "mybranch"
  retention_unit  = "d"
  retention_value = 7.0
}
```

In order to prevent Terraform from deleting an existing backup and creating a new one:

1. Manually remove the existing backup from state: `terraform state rm planetscale_backup.test`
2. Update the `name` of the `planetscale_backup` to a new name.