password: fix `cidrs` create and update
1. add `cidrs` to api request in Create()
2. add `cidrs` to api request in Update()
3. Added many test cases for updating `cidrs`

NOTE: still TODO: the user may specify a cidr range without a mask,
eg: `127.0.0.1`. The API will return `127.0.0.1/32` which
will cause terraform to error. Need to figure out the best
way to handle this.

Unrelated to the new `cidrs` field, the following changes were also made:

- The password resource had `name` as an updatable field but it would
fail because the plaintext password is not returned. Re-use the existing
value from state to work around this.

- The computed/readonly `database_branch` attribute was being marked
  as 'force replacement' every time the password was updated. Removed
  the requires-replacement marker from this attribute.
- Something is causing the password update plan to set all computed
  values to unknown, including `id`, which would break Update because that
  field is required for the API client. Set this field to `UseStateForUnknown`
  to work around.