warn for upcoming v1.0
This will be the last version in the experimental v0.x series.

There are no changes in this release other than this warning printed
at runtime:

```
╷
│ Warning: PlanetScale Provider v0.x Deprecation Notice
│
│   with provider["registry.terraform.io/planetscale/planetscale"],
│   on main.tf line 9, in provider "planetscale":
│    9: provider "planetscale" {
│
│ The v0.x versions of the PlanetScale Terraform provider are deprecated and will not receive updates. A rewrite is in progress for v1.0. Please pin your provider version and monitor the repository for the v1.0 release. The new v1.0 versions will not be backwards compatible with v0.x, so please plan your migration accordingly.`
```

A major rewrite is in progress and will be released as v1.x. The new version
will support Vitess as well as Postgres databases, including long requested
support for Vitess password and Postgres role management.

Existing PlanetScale resources managed by terraform will need to be imported into
a new state file for the v1.x release.